eldercare-platform/services/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
python-multipart==0.0.17
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
pandas==2.3.3
